move QdrantClientConfig and related structs/traits to config.rs
